Write submit function that creates message object inside onSubmit
