+++ pages/api/messages.jsx
@@ -7,7 +7,7 @@ export default async function handler(req, res) {
 
   try {
     const message = JSON.parse(req.body);
-    const savedMessage = await prisma.message.create({ data: message });
+    const savedMessage = await prisma.messages.create({ data: message });
     res.status(200).json(savedMessage);
   } catch {
     res.status(400).json({ error: "failed to load data" });
