+++ pages/index.jsx
@@ -1,8 +1,28 @@
 import Head from "next/head";
 import Image from "next/image";
 import styles from "../styles/Home.module.css";
+import { useEffect, useState } from "react";
+import io from "socket.io-client";
+let socket;
 
 export default function Home() {
+  const [input, setInput] = useState("");
+
+  useEffect(() => socketInitializer(), []);
+
+  const socketInitializer = async () => {
+    await fetch("/api/socket");
+    socket = io();
+
+    socket.on("connect", () => {
+      console.log("connected");
+    });
+
+    socket.on("update-input", msg => {
+      setInput(msg);
+      console.log("changed");
+    });
+  };
 
   return (
     <div className={styles.container}>
@@ -14,7 +34,7 @@ export default function Home() {
 
       <main className={styles.main}>
         <h1 className={styles.title}>Hello</h1>
-        
+
         <p className={styles.description}>
           Get started by editing{" "}
           <code className={styles.code}>pages/index.js</code>
